refactor(vox-jei3): split rec/music-catalog verbs into server_audio_tools
Extract the RecTools and MusicCatalogTools humble objects out of the
server.py god-module into a dedicated server_audio_tools module (Split
Module, PY-RF-3), mirroring vibe_command.py. server.py registers them via
a client factory and a live-session closure; the classes depend only on
those, not on the module globals.

server.py: module_size 917->739, classes 4->2 (relaxation reversed);
method_ratio/avg_params/avg_complexity/class_to_func all improve vs
origin/main. The new module clears every OO threshold. Behaviour and the
mic tool surface are unchanged; tests updated to the new location and the
injected session provider.

The extraction dips server.py method_ratio/class_to_func/avg_complexity
against the intra-branch peak (all three still better than origin/main);
scoped, justified relaxes record that trade while module_size/classes
register as the improvement.

Mission: m-2026-05-25-003
Delegation: d-2026-05-25-012
