+++ .oo-audit.jsonl
@@ -492,3 +492,7 @@
 {"ts":"2026-07-24T06:05:28Z","commit":"b0af958","source":null,"files_scored":168,"files_improved":1,"files_regressed":0,"verdict":"pass","reason":null,"deltas":{"src/punt_vox/cli_rec.py":{"method_ratio":[0.0,0.957],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.61],"max_complexity":[0.0,5.0],"avg_complexity":[0.0,1.78],"module_size":[0.0,287.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,0.667],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
 {"ts":"2026-07-24T06:12:25Z","commit":"494b95e","source":"vox-jei3","files_scored":1,"files_improved":0,"files_regressed":1,"verdict":"relaxed","reason":"vox-jei3 CLI: music new/get/remove catalog-authoring verbs + CatalogGateway/ClientCatalogGateway seam (mirrors ProgramGateway); module_size, classes_per_module, avg_params grow to carry the feature while method_ratio/class_to_func_ratio/avg_complexity improve","deltas":{"src/punt_vox/cli_music.py":{"avg_params":[1.0,1.25],"module_size":[144.0,232.0],"classes_per_module":[1.0,2.0]}}}
 {"ts":"2026-07-24T06:50:49Z","commit":"93dff34","source":"vox-jei3","files_scored":1,"files_improved":0,"files_regressed":1,"verdict":"relaxed","reason":"MCP D-7 parity: 8 new rec/catalog verbs + their two humble-object homes (_RecTools/_MusicCatalogTools) grow module_size and classes_per_module; the OO-correct split into a dedicated tools module is outside this mission's server.py-only write-set. method_ratio, class_to_func_ratio, avg_params, avg_complexity all improved.","deltas":{"src/punt_vox/server.py":{"module_size":[807.0,917.0],"classes_per_module":[2.0,4.0]}}}
+{"ts":"2026-07-24T07:39:46Z","commit":"34ca72d","source":null,"files_scored":169,"files_improved":0,"files_regressed":1,"verdict":"fail","reason":null,"deltas":{}}
+{"ts":"2026-07-24T07:43:41Z","commit":"34ca72d","source":null,"files_scored":1,"files_improved":0,"files_regressed":1,"verdict":"relaxed","reason":"split RecTools/MusicCatalogTools out of the server.py god-module: removing two classes (10 methods) dips method_ratio 0.708->0.63, class_to_func 0.222->0.12, avg_complexity 2.56->2.59 vs the intra-branch peak, while module_size 917->739 and classes 4->2 improve; all three dipped metrics remain better than origin/main baseline (0.615/0.118/2.66)","deltas":{"src/punt_vox/server.py":{"method_ratio":[0.708,0.632],"avg_complexity":[2.56,2.59],"class_to_func_ratio":[0.222,0.125]}}}
+{"ts":"2026-07-24T07:43:44Z","commit":"34ca72d","source":null,"files_scored":169,"files_improved":0,"files_regressed":0,"verdict":"pass","reason":null,"deltas":{}}
+{"ts":"2026-07-24T07:43:58Z","commit":"34ca72d","source":null,"files_scored":169,"files_improved":1,"files_regressed":0,"verdict":"pass","reason":null,"deltas":{"src/punt_vox/server_audio_tools.py":{"method_ratio":[0.0,0.938],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,1.41],"max_complexity":[0.0,7.0],"avg_complexity":[0.0,1.82],"module_size":[0.0,237.0],"classes_per_module":[0.0,2.0],"class_to_func_ratio":[0.0,0.667],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
