+++ .oo-coupling-audit.jsonl
@@ -145,3 +145,5 @@
 {"ts":"2026-07-24T04:50:47Z","commit":"9ee8673","source":null,"files_scored":167,"files_improved":1,"files_regressed":0,"verdict":"pass","reason":null,"deltas":{"src/punt_vox/voxd/rec_handlers.py":{"efferent_coupling":[0.0,2.0],"public_names":[0.0,2.0],"circular_imports":[0.0,0.0],"max_lcom":[0.0,0.0],"avg_lcom":[0.0,0.0]}}}
 {"ts":"2026-07-24T05:11:38Z","commit":"de855d3","source":null,"files_scored":167,"files_improved":0,"files_regressed":1,"verdict":"relaxed","reason":"5 new public catalog methods + RecordingSummary raise public_names; the stateless music_get/fetch verifier staticmethods lower LCOM (vox-jei3)","deltas":{"src/punt_vox/client.py":{"public_names":[6.0,7.0],"avg_lcom":[0.652,0.654]}}}
 {"ts":"2026-07-24T06:14:22Z","commit":"494b95e","source":null,"files_scored":168,"files_improved":4,"files_regressed":0,"verdict":"pass","reason":null,"deltas":{"src/punt_vox/__main__.py":{"efferent_coupling":[14.0,13.0],"public_names":[48.0,44.0]},"src/punt_vox/cli_rec.py":{"efferent_coupling":[0.0,5.0],"public_names":[0.0,2.0],"circular_imports":[0.0,0.0],"max_lcom":[0.0,0.286],"avg_lcom":[0.0,0.143]},"src/punt_vox/client_sync.py":{"max_lcom":[0.233,0.181],"avg_lcom":[0.233,0.181]},"src/punt_vox/types_audio.py":{"public_names":[4.0,3.0]}}}
+{"ts":"2026-07-24T07:46:16Z","commit":"34ca72d","source":null,"files_scored":169,"files_improved":1,"files_regressed":1,"verdict":"fail","reason":null,"deltas":{"src/punt_vox/server_audio_tools.py":{"efferent_coupling":[0.0,3.0],"public_names":[0.0,3.0],"circular_imports":[0.0,0.0],"max_lcom":[0.0,0.0],"avg_lcom":[0.0,0.0]}}}
+{"ts":"2026-07-24T07:46:23Z","commit":"34ca72d","source":null,"files_scored":169,"files_improved":0,"files_regressed":1,"verdict":"relaxed","reason":"extraction adds one internal import (server_audio_tools) to server.py so it can register the RecTools/MusicCatalogTools verbs; offset by the module-size/class-count reduction of the same split","deltas":{"src/punt_vox/server.py":{"efferent_coupling":[14.0,15.0]}}}
