+++ .oo-coupling-baseline.json
@@ -357,11 +357,18 @@
     "avg_lcom": 0.0
   },
   "src/punt_vox/server.py": {
-    "efferent_coupling": 14.0,
-    "public_names": 15.0,
+    "efferent_coupling": 15.0,
+    "public_names": 14.0,
     "circular_imports": 0.0,
     "max_lcom": 0.748,
     "avg_lcom": 0.748
+  },
+  "src/punt_vox/server_audio_tools.py": {
+    "efferent_coupling": 3.0,
+    "public_names": 3.0,
+    "circular_imports": 0.0,
+    "max_lcom": 0.0,
+    "avg_lcom": 0.0
   },
   "src/punt_vox/service/__init__.py": {
     "efferent_coupling": 5.0,
